skip cancelling rolling upgrades if latest rolling upgrade status is marked as completed
